requirements.txt
streamlit>=1.30.0
pillow>=10.0.0
tqdm>=4.66.0